new branch, creation of a back office, removal of bootstrap, switch from JQuery to Javascript and more
Creation of a Sanity back office so that files and data become dynamic and can be modified via an administrator interface rather than having a static page.
Removed bootstrap for performance reasons.
Removal of JQuery and its dependencies to lighten and modernize the code.
Refactoring of the gallery and lightbox in JavaScript.